requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
lifelines>=0.28
plotly>=5.18
streamlit>=1.30
anthropic>=0.40
python-dotenv>=1.0
requests>=2.31
pyarrow>=14.0
scipy>=1.11
shap>=0.44
pytest>=7.4
